Cleaned up the constants to separate file
